refactor: 重构并工程化用户交互相关接口
